refactor: Authentication 도메인 DTO 패키지 위치수정
