+++ src/main/java/com/school/mohitto/domain/authentication/dto/LoginWithAuthorizationCodeRequest.java
@@ -1,4 +1,4 @@
-package com.school.mohitto.domain.authentication;
+package com.school.mohitto.domain.authentication.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
